+++ src/jsonrpc-machine.hpp
@@ -153,6 +153,7 @@ private:
     uarch_cycle_root_hashes do_collect_uarch_cycle_root_hashes(uint64_t mcycle_end,
         int32_t log2_bundle_uarch_cycle_count, const machine_hashes &revert_uarch_tail) override;
     address_range_descriptions do_get_address_ranges() const override;
+    machine_cmio_request do_receive_cmio_request(std::span<uint8_t> data) const override;
     void do_send_cmio_response(const_machine_hash_view revert_root_hash, uint16_t reason, const unsigned char *data,
         uint64_t length) override;
     access_log do_log_send_cmio_response(const_machine_hash_view revert_root_hash, uint16_t reason,